Passing all Authorization unit tests
