+++ libraries/microsoft-agents-hosting-core/tests/tools/mock_user_token_client.py
